Simplify page cache adversarial tests, reorganise page cache harness tests
Remove BootClassPathRunner infrustructure of page cache adversarial tests since those where extremelly flaky under IBM JDK.
Switch those tests to inject and use adversarial file system or adversarial channel instead of custom file dispatcher.

Extract page cache harness into separate tests.
