+++ community/io/src/test/java/org/neo4j/io/pagecache/PageCacheTest.java
@@ -79,6 +79,7 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
+import static org.junit.Assume.assumeTrue;
 import static org.neo4j.io.pagecache.PagedFile.PF_NO_FAULT;
 import static org.neo4j.io.pagecache.PagedFile.PF_NO_GROW;
 import static org.neo4j.io.pagecache.PagedFile.PF_SHARED_READ_LOCK;
@@ -1563,24 +1564,6 @@ public abstract class PageCacheTest<T extends PageCache> extends PageCacheTestSu
             }
         }
         pagedFile.close();
-    }
-
-    @RepeatRule.Repeat( times = 10 )
-    @Test( timeout = SEMI_LONG_TIMEOUT_MILLIS )
-    public void readsAndWritesMustBeMutuallyConsistent() throws Exception
-    {
-        int filePageCount = 100;
-        RandomPageCacheTestHarness harness = new RandomPageCacheTestHarness();
-        harness.disableCommands( FlushCache, FlushFile, MapFile, UnmapFile );
-        harness.setCommandProbabilityFactor( ReadRecord, 0.5 );
-        harness.setCommandProbabilityFactor( WriteRecord, 0.5 );
-        harness.setConcurrencyLevel( 8 );
-        harness.setFilePageCount( filePageCount );
-        harness.setInitialMappedFiles( 1 );
-        harness.setCachePageSize( pageCachePageSize );
-        harness.setFilePageSize( pageCachePageSize );
-        harness.setVerification( filesAreCorrectlyWrittenVerification( new StandardRecordFormat(), filePageCount ) );
-        harness.run( SEMI_LONG_TIMEOUT_MILLIS, TimeUnit.MILLISECONDS );
     }
 
     @Test( timeout = SHORT_TIMEOUT_MILLIS )
@@ -3172,137 +3155,6 @@ public abstract class PageCacheTest<T extends PageCache> extends PageCacheTestSu
 
         pagedFileA.close();
         pagedFileB.close();
-    }
-
-    @Test( timeout = SEMI_LONG_TIMEOUT_MILLIS )
-    public void concurrentPageFaultingMustNotPutInterleavedDataIntoPages() throws Exception
-    {
-        final int filePageCount = 11;
-        final RecordFormat recordFormat = new PageCountRecordFormat();
-        RandomPageCacheTestHarness harness = new RandomPageCacheTestHarness();
-        harness.setConcurrencyLevel( 11 );
-        harness.setUseAdversarialIO( false );
-        harness.setCachePageCount( 3 );
-        harness.setCachePageSize( pageCachePageSize );
-        harness.setFilePageCount( filePageCount );
-        harness.setFilePageSize( pageCachePageSize );
-        harness.setInitialMappedFiles( 1 );
-        harness.setCommandCount( 10000 );
-        harness.setRecordFormat( recordFormat );
-        harness.setFileSystem( fs );
-        harness.disableCommands( FlushCache, FlushFile, MapFile, UnmapFile, WriteRecord, WriteMulti );
-        harness.setPreparation( ( pageCache1, fs1, filesTouched ) -> {
-            File file = filesTouched.iterator().next();
-            try ( PagedFile pf = pageCache1.map( file, pageCachePageSize );
-                  PageCursor cursor = pf.io( 0, PF_SHARED_WRITE_LOCK ) )
-            {
-                for ( int pageId = 0; pageId < filePageCount; pageId++ )
-                {
-                    cursor.next();
-                    recordFormat.fillWithRecords( cursor );
-                }
-            }
-        } );
-
-        harness.run( SEMI_LONG_TIMEOUT_MILLIS, MILLISECONDS );
-    }
-
-    @Test( timeout = SEMI_LONG_TIMEOUT_MILLIS )
-    public void concurrentFlushingMustNotPutInterleavedDataIntoFile() throws Exception
-    {
-        final RecordFormat recordFormat = new StandardRecordFormat();
-        final int filePageCount = 2_000;
-        RandomPageCacheTestHarness harness = new RandomPageCacheTestHarness();
-        harness.setConcurrencyLevel( 16 );
-        harness.setUseAdversarialIO( false );
-        harness.setCachePageCount( filePageCount / 2 );
-        harness.setFilePageCount( filePageCount );
-        harness.setCachePageSize( pageCachePageSize );
-        harness.setFilePageSize( pageCachePageSize );
-        harness.setInitialMappedFiles( 3 );
-        harness.setCommandCount( 15_000 );
-        harness.setFileSystem( fs );
-        harness.disableCommands( MapFile, UnmapFile, ReadRecord, ReadMulti );
-        harness.setVerification( filesAreCorrectlyWrittenVerification( recordFormat, filePageCount ) );
-
-        harness.run( SEMI_LONG_TIMEOUT_MILLIS, MILLISECONDS );
-    }
-
-    @Test( timeout = SEMI_LONG_TIMEOUT_MILLIS )
-    public void concurrentFlushingWithMischiefMustNotPutInterleavedDataIntoFile() throws Exception
-    {
-        final RecordFormat recordFormat = new StandardRecordFormat();
-        final int filePageCount = 2_000;
-        RandomPageCacheTestHarness harness = new RandomPageCacheTestHarness();
-        harness.setConcurrencyLevel( 16 );
-        harness.setUseAdversarialIO( true );
-        harness.setMischiefRate( 0.5 );
-        harness.setFailureRate( 0.0 );
-        harness.setErrorRate( 0.0 );
-        harness.setCachePageCount( filePageCount / 2 );
-        harness.setFilePageCount( filePageCount );
-        harness.setCachePageSize( pageCachePageSize );
-        harness.setFilePageSize( pageCachePageSize );
-        harness.setInitialMappedFiles( 3 );
-        harness.setCommandCount( 15_000 );
-        harness.setFileSystem( fs );
-        harness.disableCommands( MapFile, UnmapFile, ReadRecord, ReadMulti );
-        harness.setVerification( filesAreCorrectlyWrittenVerification( recordFormat, filePageCount ) );
-
-        harness.run( SEMI_LONG_TIMEOUT_MILLIS, MILLISECONDS );
-    }
-
-    @Test( timeout = SEMI_LONG_TIMEOUT_MILLIS )
-    public void concurrentFlushingWithFailuresMustNotPutInterleavedDataIntoFile() throws Exception
-    {
-        final RecordFormat recordFormat = new StandardRecordFormat();
-        final int filePageCount = 20_000;
-        RandomPageCacheTestHarness harness = new RandomPageCacheTestHarness();
-        harness.setConcurrencyLevel( 16 );
-        harness.setUseAdversarialIO( true );
-        harness.setMischiefRate( 0.0 );
-        harness.setFailureRate( 0.5 );
-        harness.setErrorRate( 0.0 );
-        harness.setCachePageCount( filePageCount / 2 );
-        harness.setFilePageCount( filePageCount );
-        harness.setCachePageSize( pageCachePageSize );
-        harness.setFilePageSize( pageCachePageSize );
-        harness.setInitialMappedFiles( 3 );
-        harness.setCommandCount( 150_000 );
-        harness.setFileSystem( fs );
-        harness.disableCommands( MapFile, UnmapFile, ReadRecord, ReadMulti );
-        harness.setVerification( filesAreCorrectlyWrittenVerification( recordFormat, filePageCount ) );
-
-        harness.run( SEMI_LONG_TIMEOUT_MILLIS, MILLISECONDS );
-    }
-
-    private Phase filesAreCorrectlyWrittenVerification( final RecordFormat recordFormat, final int filePageCount )
-    {
-        return ( pageCache1, fs1, filesTouched ) -> {
-            for ( File file : filesTouched )
-            {
-                try ( PagedFile pf = pageCache1.map( file, pageCachePageSize );
-                      PageCursor cursor = pf.io( 0, PF_SHARED_READ_LOCK ) )
-                {
-                    for ( int pageId = 0; pageId < filePageCount && cursor.next(); pageId++ )
-                    {
-                        try
-                        {
-                            recordFormat.assertRecordsWrittenCorrectly( cursor );
-                        }
-                        catch ( Throwable th )
-                        {
-                            th.addSuppressed( new Exception( "pageId = " + pageId ) );
-                            throw th;
-                        }
-                    }
-                }
-                try ( StoreChannel channel = fs1.open( file, "r" ) )
-                {
-                    recordFormat.assertRecordsWrittenCorrectly( file, channel );
-                }
-            }
-        };
     }
 
     @Test( timeout = SEMI_LONG_TIMEOUT_MILLIS )
